feat: test github labels
