+++ .github/workflows/stale.yml
@@ -26,11 +26,11 @@ jobs:
       with:
         repo-token: ${{ secrets.BOT_GITHUB_TOKEN }}
         days-before-stale: 7
-        days-before-close: 5
+        days-before-close: 3
         stale-issue-message: 'This issue is stale because it has been open 7 days with no activity. Remove stale label or comment or this will be closed in 7 days.'
         stale-pr-message: 'This issue is stale because it has been open 7 days with no activity.'
-        close-issue-message: 'This issue was closed because it has been stalled for 5 days with no activity.'
-        close-pr-message: 'This PR was closed because it has been stalled for 5 days with no activity. You can reopen it if you want.'
+        close-issue-message: 'This issue was closed because it has been stalled for 3 days with no activity.'
+        close-pr-message: 'This PR was closed because it has been stalled for 3 days with no activity. You can reopen it if you want.'
         stale-pr-label: lifecycle/stale
         stale-issue-label: lifecycle/stale
         exempt-issue-labels: 'kubecub'
