+++ .github/workflows/test.yml
@@ -15,9 +15,9 @@ jobs:
         uses: actions/checkout@v2
       
       - name: Install dependencies
-        run: sudo apt-get update && sudo apt-get install -y make
+        run: sudo apt-get update 
+      - name: Build project&& sudo apt-get install -y make
       
-      - name: Build project
         run: make build
 
       - name: Test build artifacts
